requirements.txt
pandas==2.2.1
requests==2.31.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.1.2
Pillow==10.3.0